LPS-183059 Rename packages
